requirements.txt
langchain-groq==0.2.5
langchain-core==0.3.44
langgraph==0.3.7
python-dotenv==1.0.1
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6